requirements.txt
pymongo
pyTenable
defusedxml

